requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
boto==2.49.0
boto3==1.24.29
botocore==1.27.29
celery==5.2.7
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.7
django-admin-rangefilter==0.8.1
django-allauth==0.51.0
django-appconf==1.0.5
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-compressor==4.0
django-crispy-forms==1.12.0
django-htmlmin==0.11.0
django-storages==1.12.3
django-storages-redux==1.3.3
django-timezone-field==5.0
docutils==0.17.1
gunicorn==20.1.0
html5lib==1.1
idna==3.2
jmespath==1.0.0
kombu==5.2.4
oauthlib==3.1.1
packaging==21.3
Pillow==8.3.2
prompt-toolkit==3.0.30
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.20
PyJWT==2.4.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-memcached==1.59
python3-openid==3.2.0
pytz==2022.1
rcssmin==1.1.0
redis==4.3.4
requests==2.26.0
requests-oauthlib==1.3.1
rjsmin==1.2.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
suit==2.0.2
toml==0.10.2
twilio==7.9.3
tzdata==2022.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
